requirements.txt
Django==5.2.4
djangorestframework==3.15.2
Pillow==10.4.0
django-cors-headers==4.3.1
